drivers: remove mlx* glue libraries separate ABI version
The glue libraries are tightly bound to the mlx drivers of a dpdk version
and are packaged with them.

Keeping a separate ABI version prevents us from installing two versions of
dpdk.
Maintaining this separate version just adds confusion.
Align the glue library ABI version to the global ABI version.

Signed-off-by: David Marchand <david.marchand@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>