doc: change references to blacklist and whitelist
There are two areas where documentation needed update.
The first was use of whitelist when describing address
filtering.

The other is the legacy -w whitelist option for PCI
which is used in many examples

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Acked-by: Luca Boccassi <bluca@debian.org>

doc: replace -w with -a in the documentation

The -w option is deprecated and replaced with -a

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>