eal: create a runtime directory on windows
Added eal_create_runtime_dir() function in order to run any application
as a user that does not have administrator access.
Currently, since there is no runtime directory set, the code tries to
create a file in C:\ which is only writable with administrator
privileges. As a result, if the user is not admin, the application will
fail.

Signed-off-by: Ranjit Menon <ranjit.menon@intel.com>
Signed-off-by: Pallavi Kadam <pallavi.kadam@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>