net/vdev_netvsc: prevent alarm loss on failed device probe
If a device probe fails, the alarm is canceled and will no longer work for
previously probed devices.

Fix this by checking if alarm is necessary at the end of each device probe.
Reset the alarm if there are vdev_netvsc_ctx created.

Change log:
v2: removed lock and flags, use counter to decide if alarm should be reset

Cc: stable@dpdk.org
Signed-off-by: Long Li <longli@microsoft.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>